# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "flight-review-server",
    "description": "Flight Review v2 — HTTP API server",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/api/logs": {
      "get": {
        "tags": [
          "Logs"
        ],
        "summary": "GET /api/logs -- list with filters",
        "operationId": "list_logs",
        "parameters": [
          {
            "name": "sys_name",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ver_hw",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "include_private",
            "in": "query",
            "description": "If true, include private logs in results (default: only public)",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "date_from",
            "in": "query",
            "description": "ISO-8601 date string lower bound on created_at",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "date_to",
            "in": "query",
            "description": "ISO-8601 date string upper bound on created_at",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "flight_duration_min",
            "in": "query",
            "description": "Minimum flight duration in seconds",
            "required": false,
            "schema": {
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "flight_duration_max",
            "in": "query",
            "description": "Maximum flight duration in seconds",
            "required": false,
            "schema": {
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "ver_sw_release_str",
            "in": "query",
            "description": "Filter by software release version string (prefix match)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ver_sw",
            "in": "query",
            "description": "Filter by software git hash (exact match)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sys_uuid",
            "in": "query",
            "description": "Filter by system UUID (exact match)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "vehicle_type",
            "in": "query",
            "description": "Filter by vehicle type category (exact match)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "localization",
            "in": "query",
            "description": "Filter by localization source (substring match)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "vibration_status",
            "in": "query",
            "description": "Filter by vibration status (exact match)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "has_gps",
            "in": "query",
            "description": "Filter by GPS presence (lat IS NOT NULL)",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "location_name",
            "in": "query",
            "description": "Filter by location name (substring match)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Sort column and direction, e.g. \"created_at:desc\", \"flight_duration_s:asc\"",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "has_topic",
            "in": "query",
            "description": "Filter logs that contain this topic name",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "parameter",
            "in": "query",
            "description": "Filter by parameter \"name:value\" (e.g. \"EKF2_AID_MASK:24\")",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tag",
            "in": "query",
            "description": "Filter by tag",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "error_message",
            "in": "query",
            "description": "Filter by error message substring",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "lat",
            "in": "query",
            "description": "Center latitude for geographic search",
            "required": false,
            "schema": {
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "lon",
            "in": "query",
            "description": "Center longitude for geographic search",
            "required": false,
            "schema": {
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "radius_km",
            "in": "query",
            "description": "Radius in km for geographic search",
            "required": false,
            "schema": {
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "field_max",
            "in": "query",
            "description": "Filter: field max exceeds threshold. Format: \"topic.field:value\"",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "field_min",
            "in": "query",
            "description": "Filter: field min below threshold. Format: \"topic.field:value\"",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "diagnostic",
            "in": "query",
            "description": "Filter by diagnostic presence (e.g. \"motor_failure\")",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "diagnostic_severity",
            "in": "query",
            "description": "Filter by diagnostic severity (e.g. \"critical\")",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Filtered and paginated log list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/logs/{id}": {
      "get": {
        "tags": [
          "Logs"
        ],
        "summary": "GET /api/logs/:id -- single log metadata",
        "operationId": "get_log",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Log UUID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Log metadata",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogRecord"
                }
              }
            }
          },
          "404": {
            "description": "Log not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Logs"
        ],
        "summary": "DELETE /api/logs/:id?token=<delete_token>",
        "operationId": "delete_log",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Log UUID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "token",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Log deleted"
          },
          "403": {
            "description": "Invalid delete token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Log not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/logs/{id}/data/{filename}": {
      "get": {
        "tags": [
          "Logs"
        ],
        "summary": "GET /api/logs/:id/data/:filename -- serve Parquet/metadata files\nSupports HTTP Range requests for DuckDB-WASM compatibility.\nIf the file does not exist in v2 storage but a v1 ULG prefix is configured,\nattempts lazy conversion of the original .ulg file on first access.",
        "operationId": "get_log_file",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Log UUID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "filename",
            "in": "path",
            "description": "Stored file name, e.g. metadata.json or a Parquet file",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Stored JSON or binary log artifact",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "integer",
                    "format": "int32",
                    "minimum": 0
                  }
                }
              }
            }
          },
          "206": {
            "description": "Partial content for range requests",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "integer",
                    "format": "int32",
                    "minimum": 0
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid range request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "File not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/logs/{id}/track": {
      "get": {
        "tags": [
          "Logs"
        ],
        "summary": "GET /api/logs/:id/track -- lightweight GPS track for thumbnails\nReturns a downsampled GPS track (max 100 points) for canvas rendering.\nMuch smaller than fetching the full metadata.json (~2KB vs ~200KB).",
        "operationId": "get_track",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Log UUID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Downsampled GPS track",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TrackPointCompact"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Log data not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
      "ListResponse": {
        "type": "object",
        "required": [
          "logs",
          "total"
        ],
        "properties": {
          "logs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LogRecord"
            }
          },
          "total": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "TrackPointCompact": {
        "type": "object",
        "required": [
          "lat",
          "lon",
          "m"
        ],
        "properties": {
          "lat": {
            "type": "number",
            "format": "double"
          },
          "lon": {
            "type": "number",
            "format": "double"
          },
          "m": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        }
      },
...
    }
  },
...
}